# api/openapi.yaml
openapi: 3.1.0
info:
  title: Open Source Project Intelligence API
  version: 1.0.0
  description: Repository-owned frozen HTTP contract. Resource IDs are opaque decimal strings.
servers:
  - url: http://localhost:8100
paths:
...
  /api/v1/catalog/projects:
    get:
      operationId: getApiV1CatalogProjects
      tags: [contract]
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/catalog/projects/{project_id}:
    get:
      operationId: getApiV1CatalogProjectsProjectId
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/session:
    get:
      operationId: getApiV1Session
      tags: [contract]
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/portfolio:
    get:
      operationId: getApiV1Portfolio
      tags: [contract]
      parameters:
        - name: window
          in: query
          required: false
          schema:
            type: string
            description: Preset 30d, 90d, 180d, 365d or half-open YYYY-MM-DD/YYYY-MM-DD interval.
        - name: cutoff
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects:
    get:
      operationId: getApiV1Projects
      tags: [contract]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: postApiV1Projects
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}:
    get:
      operationId: getApiV1ProjectsProjectId
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchApiV1ProjectsProjectId
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/jobs:
    get:
      operationId: getApiV1ProjectsProjectIdJobs
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: kind
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/contributors:
    get:
      operationId: getApiV1ProjectsProjectIdContributors
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/adoption:
    get:
      operationId: getApiV1ProjectsProjectIdAdoption
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/security:
    get:
      operationId: getApiV1ProjectsProjectIdSecurity
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/comparisons:
    post:
      operationId: postApiV1Comparisons
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/trends:
    get:
      operationId: getApiV1ProjectsProjectIdTrends
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: kind
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/recommendation:
    get:
      operationId: getApiV1ProjectsProjectIdRecommendation
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: policy
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cutoff
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/policies:
    get:
      operationId: getApiV1Policies
      tags: [contract]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: postApiV1Policies
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/radar:
    get:
      operationId: getApiV1Radar
      tags: [contract]
      parameters:
        - name: policy
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/topics:
    get:
      operationId: getApiV1ProjectsProjectIdTopics
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/releases:
    get:
      operationId: getApiV1ProjectsProjectIdReleases
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/projects/{project_id}/crawls:
    post:
      operationId: postApiV1ProjectsProjectIdCrawls
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/knowledge/search:
    post:
      operationId: postApiV1ProjectsProjectIdKnowledgeSearch
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/queries:
    post:
      operationId: postApiV1ProjectsProjectIdQueries
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/alerts:
    get:
      operationId: getApiV1Alerts
      tags: [contract]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: project
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/alert-rules:
    post:
      operationId: postApiV1AlertRules
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  schemas:
...
    Document:
      type: object
      additionalProperties: true
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, pattern: '^[a-z][a-z0-9_]*$' }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: '^[a-z][a-z0-9_]*$' }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
  responses:
    Problem:
      description: A safe RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'